"Update Posts and ReactionsButtons components to use Redux for state management and dispatch reactions"
This diff updates the Posts and ReactionsButtons components to use Redux for state management and dispatch reactions. The Posts component now passes the entire post object to the ReactionsButtons component, which then uses the post ID to dispatch the addReaction action. The ReactionsButtons component has also been updated to use the useDispatch hook from react-redux and to display the number of reactions for each emoji. Additionally, the postSlice has been updated to include an addReaction action that increments the number of reactions for a specific post and reaction.